# Add doc-values aggregators for the array and collect set aggregations.