--- a/server/src/main/java/io/crate/execution/engine/aggregation/impl/ArrayAgg.java
+++ b/server/src/main/java/io/crate/execution/engine/aggregation/impl/ArrayAgg.java
@@ -27,12 +27,32 @@ import io.crate.breaker.SizeEstimator;
 import io.crate.breaker.SizeEstimatorFactory;
 import io.crate.data.Input;
 import io.crate.execution.engine.aggregation.AggregationFunction;
+import io.crate.execution.engine.aggregation.DocValueAggregator;
 import io.crate.memory.MemoryManager;
 import io.crate.metadata.functions.Signature;
+import io.crate.types.ByteType;
 import io.crate.types.DataType;
+import io.crate.types.DoubleType;
+import io.crate.types.FloatType;
+import io.crate.types.IntegerType;
+import io.crate.types.IpType;
+import io.crate.types.LongType;
+import io.crate.types.ShortType;
+import io.crate.types.StringType;
+import io.crate.types.TimestampType;
+import org.apache.lucene.index.DocValues;
+import org.apache.lucene.index.LeafReader;
+import org.apache.lucene.index.SortedNumericDocValues;
+import org.apache.lucene.util.NumericUtils;
+import org.apache.lucene.util.RamUsageEstimator;
 import org.elasticsearch.Version;
 import org.elasticsearch.common.breaker.CircuitBreakingException;
+import org.elasticsearch.index.fielddata.FieldData;
+import org.elasticsearch.index.fielddata.SortedBinaryDocValues;
+import org.elasticsearch.index.mapper.MappedFieldType;
 
+import javax.annotation.Nullable;
+import java.io.IOException;
 import java.util.ArrayList;
 import java.util.List;
 
@@ -107,5 +127,229 @@ public final class ArrayAgg extends AggregationFunction<List<Object>, List<Objec
     @Override
     public DataType<?> partialType() {
         return boundSignature.getReturnType().createType();
+    }
+
+    @Nullable
+    @Override
+    public DocValueAggregator<?> getDocValueAggregator(List<DataType<?>> argumentTypes,
+                                                       List<MappedFieldType> fieldTypes) {
+        var dataType = argumentTypes.get(0);
+        switch (dataType.id()) {
+            case ByteType.ID:
+            case ShortType.ID:
+            case IntegerType.ID:
+            case LongType.ID:
+            case TimestampType.ID_WITH_TZ:
+            case TimestampType.ID_WITHOUT_TZ:
+                return new LongArrayAggNumericDocValueAggregator(
+                    fieldTypes.get(0).name(),
+                    dataType,
+                    sizeEstimator
+                );
+            case FloatType.ID:
+                return new FloatArrayAggNumericDocValueAggregator(
+                    fieldTypes.get(0).name(),
+                    sizeEstimator
+                );
+            case DoubleType.ID:
+                return new DoubleArrayAggNumericDocValueAggregator(
+                    fieldTypes.get(0).name(),
+                    sizeEstimator
+                );
+            case IpType.ID:
+            case StringType.ID:
+                return new ArrayAggBinaryDocValueAggregator(
+                    fieldTypes.get(0).name(),
+                    sizeEstimator
+                );
+            default:
+                return null;
+        }
+    }
+
+    private static class LongArrayAggNumericDocValueAggregator implements DocValueAggregator<ArrayList<Object>> {
+
+        private final String columnName;
+        private final DataType<?> dataType;
+        private final SizeEstimator<Object> sizeEstimator;
+
+        private SortedNumericDocValues values;
+
+        public LongArrayAggNumericDocValueAggregator(String columnName,
+                                                     DataType<?> dataType,
+                                                     SizeEstimator<Object> sizeEstimator) {
+            this.columnName = columnName;
+            this.dataType = dataType;
+            this.sizeEstimator = sizeEstimator;
+        }
+
+        @Override
+        public ArrayList<Object> initialState(RamAccounting ramAccounting) {
+            var state = new ArrayList<>();
+            ramAccounting.addBytes(RamUsageEstimator.sizeOfCollection(state));
+            return state;
+        }
+
+        @Override
+        public void loadDocValues(LeafReader reader) throws IOException {
+            values = DocValues.getSortedNumeric(reader, columnName);
+        }
+
+        @Override
+        public void apply(RamAccounting ramAccounting, int doc, ArrayList<Object> state) throws IOException {
+            if (values.advanceExact(doc)) {
+                if (values.docValueCount() == 1) {
+                    var value = dataType.sanitizeValue(values.nextValue());
+                    ramAccounting.addBytes(sizeEstimator.estimateSize(value));
+                    state.add(value);
+                }
+            } else {
+                state.add(null);
+            }
+        }
+
+        @Nullable
+        @Override
+        public Object partialResult(RamAccounting ramAccounting, ArrayList<Object> state) {
+            return state;
+        }
+    }
+
+    private static class FloatArrayAggNumericDocValueAggregator implements DocValueAggregator<ArrayList<Float>> {
+
+        private final String columnName;
+        private final SizeEstimator<Object> sizeEstimator;
+
+        private SortedNumericDocValues values;
+
+
+        public FloatArrayAggNumericDocValueAggregator(String columnName,
+                                                      SizeEstimator<Object> sizeEstimator) {
+            this.columnName = columnName;
+            this.sizeEstimator = sizeEstimator;
+        }
+
+        @Override
+        public ArrayList<Float> initialState(RamAccounting ramAccounting) {
+            var state = new ArrayList<Float>();
+            ramAccounting.addBytes(RamUsageEstimator.sizeOfCollection(state));
+            return state;
+        }
+
+        @Override
+        public void loadDocValues(LeafReader reader) throws IOException {
+            values = DocValues.getSortedNumeric(reader, columnName);
+        }
+
+        @Override
+        public void apply(RamAccounting ramAccounting, int doc, ArrayList<Float> state) throws IOException {
+            if (values.advanceExact(doc)) {
+                if (values.docValueCount() == 1) {
+                    var value = NumericUtils.sortableIntToFloat((int) values.nextValue());
+                    ramAccounting.addBytes(sizeEstimator.estimateSize(value));
+                    state.add(value);
+                }
+            } else {
+                state.add(null);
+            }
+        }
+
+        @Nullable
+        @Override
+        public Object partialResult(RamAccounting ramAccounting, ArrayList<Float> state) {
+            return state;
+        }
+    }
+
+    private static class DoubleArrayAggNumericDocValueAggregator implements DocValueAggregator<ArrayList<Double>> {
+
+        private final String columnName;
+        private final SizeEstimator<Object> sizeEstimator;
+
+        private SortedNumericDocValues values;
+
+
+        public DoubleArrayAggNumericDocValueAggregator(String columnName,
+                                                       SizeEstimator<Object> sizeEstimator) {
+            this.columnName = columnName;
+            this.sizeEstimator = sizeEstimator;
+        }
+
+        @Override
+        public ArrayList<Double> initialState(RamAccounting ramAccounting) {
+            var state = new ArrayList<Double>();
+            ramAccounting.addBytes(RamUsageEstimator.sizeOfCollection(state));
+            return state;
+        }
+
+        @Override
+        public void loadDocValues(LeafReader reader) throws IOException {
+            values = DocValues.getSortedNumeric(reader, columnName);
+        }
+
+        @Override
+        public void apply(RamAccounting ramAccounting, int doc, ArrayList<Double> state) throws IOException {
+            if (values.advanceExact(doc)) {
+                if (values.docValueCount() == 1) {
+                    var value = NumericUtils.sortableLongToDouble(values.nextValue());
+                    ramAccounting.addBytes(sizeEstimator.estimateSize(value));
+                    state.add(value);
+
+                }
+            } else {
+                state.add(null);
+            }
+        }
+
+        @Nullable
+        @Override
+        public Object partialResult(RamAccounting ramAccounting, ArrayList<Double> state) {
+            return state;
+        }
+    }
+
+    private static class ArrayAggBinaryDocValueAggregator implements DocValueAggregator<ArrayList<String>> {
+
+        private final String columnName;
+        private final SizeEstimator<Object> sizeEstimator;
+
+        private SortedBinaryDocValues values;
+
+        public ArrayAggBinaryDocValueAggregator(String columnName,
+                                                SizeEstimator<Object> sizeEstimator) {
+            this.columnName = columnName;
+            this.sizeEstimator = sizeEstimator;
+        }
+
+        @Override
+        public ArrayList<String> initialState(RamAccounting ramAccounting) {
+            var state = new ArrayList<String>();
+            ramAccounting.addBytes(RamUsageEstimator.sizeOfCollection(state));
+            return state;
+        }
+
+        @Override
+        public void loadDocValues(LeafReader reader) throws IOException {
+            values = FieldData.toString(DocValues.getSortedSet(reader, columnName));
+        }
+
+        @Override
+        public void apply(RamAccounting ramAccounting, int doc, ArrayList<String> state) throws IOException {
+            if (values.advanceExact(doc)) {
+                if (values.docValueCount() == 1) {
+                    String value = values.nextValue().utf8ToString();
+                    ramAccounting.addBytes(sizeEstimator.estimateSize(value));
+                    state.add(value);
+                }
+            } else {
+                state.add(null);
+            }
+        }
+
+        @Nullable
+        @Override
+        public Object partialResult(RamAccounting ramAccounting, ArrayList<String> state) {
+            return state;
+        }
     }
 }