… server/src/test/java/io/crate/execution/engine/aggregation/impl/CollectSetAggregationTest.java