## Summary of the changes / Why this improves CrateDB

See commits.

### array_agg
With doc values aggregator:

```
# Running setUp
# Running benchmark

## Running Query:
   Name: global array_agg on long
   Statement:
     select array_agg(duration) from uservisits
   Concurrency: 1
   Iterations: 400
100%| 400/400 [02:13<00:00,  2.99 requests/s]
Runtime (in ms):
    mean:    117.533 ± 5.925
    min/max: 70.763 → 1113.618
Percentile:
    50:   110.825 ± 60.463 (stdev)
    95:   162.934
    99.9: 1113.618

## Running Query:
   Name: global array_agg on float
   Statement:
     select array_agg("adRevenue") from uservisits
   Concurrency: 1
   Iterations: 400
100%| 400/400 [09:02<00:00,  1.36s/ requests]
Runtime (in ms):
    mean:    891.798 ± 33.724
    min/max: 346.824 → 3227.598
Percentile:
    50:   835.351 ± 344.117 (stdev)
    95:   1515.606
    99.9: 3227.598
```

Without doc values aggregator:
```
# Running setUp
# Running benchmark

## Running Query:
   Name: global array_agg on long
   Statement:
     select array_agg(duration) from uservisits
   Concurrency: 1
   Iterations: 400
100%| 400/400 [03:37<00:00,  1.84 requests/s]
Runtime (in ms):
    mean:    192.064 ± 10.599
    min/max: 94.740 → 1898.387
Percentile:
    50:   172.868 ± 108.148 (stdev)
    95:   329.987
    99.9: 1898.387

## Running Query:
   Name: global array_agg on float
   Statement:
     select array_agg("adRevenue") from uservisits
   Concurrency: 1
   Iterations: 400
100%| 400/400 [12:16<00:00,  1.84s/ requests]
Runtime (in ms):
    mean:    1181.211 ± 52.584
    min/max: 431.027 → 4477.084
Percentile:
    50:   1020.053 ± 536.576 (stdev)
    95:   2217.083
    99.9: 4477.084
```

## Checklist

 - [ ] Added an entry in `CHANGES.txt` for user facing changes
 - [ ] Updated documentation & `sql_features` table for user facing changes
 - [x] Touched code is covered by tests
 - [x] [CLA](https://crate.io/community/contribute/cla/) is signed
 - [x] This does not contain breaking changes, or if it does:
    - It is released within a major release
    - It is recorded in ``CHANGES.txt``
    - It was marked as deprecated in an earlier release if possible
    - You've thought about the consequences and other components are adapted
      (E.g. AdminUI)