@mfussenegger 2020-08-17
Given that these are now aggregations which can consume a lot of memory, we need to integrate `RamAccounting`