@kovrus 2020-08-21
@mfussenegger can you please take a look at it? I also not sure about benchmarks, the overhead probably is somewhere else, e.g. gc, in a couple of runs I am getting a bit worse results for the optimized float case.