@kovrus 2020-08-24
closing it. the doc values aggregator optimization for `collect_set`, `array_agg` and `string_agg` does not improve the performance. we tried to use as well hccp collections as the state for some aggregation which also didn't show much of improvement.